+++ CHANGELOG.md
@@ -3,6 +3,9 @@ __In descending date order first__ (most recent date and time first).
 Timestamp Format: __day month HOURS:MINUTES (military time) USER__  
 Example: 10 mar 21:13 SAJID\
 
+## 1 apr Mohammed
+* database exporting
+
 ## 13 APRIL 09:58 Jenny
 * Changed weekly schedule view in doctor to 24 hour view.
 * Changed order of right action panel.